+++ test/example_verify_test.rb
@@ -2,82 +2,19 @@
 
 require "minitest/autorun"
 require "minitest/pride"
-require "yaml"
 require "nokogiri"
 
 require_relative '../lib/rsyntaxtree'
 require_relative '../lib/rsyntaxtree/utils'
+require_relative '../dev/example_options'
 
 class ExampleParserTest < Minitest::Test
   examples_dir = File.expand_path(File.join(__dir__, "..", "docs", "_examples"))
-  svg_dir = File.expand_path(File.join(__dir__, "..", "docs", "assets", "svg"))
 
   Dir.glob("*.md", base: examples_dir).map do |md|
-    md = File.join(examples_dir, md)
-    config = YAML.load_file(md)
-    rst = File.read(md).scan(/```([^`]+)```/m).last.first
+    name, opts = ExampleOptions.load(File.join(examples_dir, md))
 
-    opts = DEFAULT_OPTS.dup
-    name = nil
-    config.each do |key, value|
-      next if value.to_s == ""
-
-      case key
-      when "name"
-        name = value
-      when "color"
-        opts[:color] = case value
-                       when "modern", "on", "true"
-                         "modern"
-                       when "traditional"
-                         "traditional"
-                       else
-                         "off"
-                       end
-      when "polyline"
-        opts[:polyline] = value
-      when "hide_default_connectors"
-        opts[:hide_default_connectors] = value
-      when "connector_height"
-        opts[:vheight] = value
-      when "symmetrization"
-        opts[:symmetrize] = value
-      when "connector"
-        opts[:leafstyle] = value
-      when "direction"
-        opts[:direction] = value
-      when "tidy"
-        opts[:tidy] = value
-      when "hspacing"
-        opts[:hspacing] = value
-      when "tidy_spacing"
-        opts[:tidy_spacing] = value
-      when "mirror"
-        opts[:mirror] = value
-      when "font"
-        opts[:fontstyle] = case value
-                           when /mono/i
-                             "mono"
-                           when /sans/i
-                             "sans"
-                           when /serif/i
-                             "serif"
-                           when /wqy|cjk/i
-                             "cjk"
-                           else
-                             "sans"
-                           end
-      end
-    end
-
-    opts[:data] = rst
-    rsg = RSyntaxTree::RSGenerator.new(opts)
-
-    #################################
-    # To test SVG, run the code below
-    #################################
-    svg = rsg.draw_svg
-
+    svg = RSyntaxTree::RSGenerator.new(opts).draw_svg
     document = Nokogiri::XML(svg)
 
     define_method "test_#{name}" do